[expo-cli] move displayLogs function to a separate module
