+++ packages/expo-cli/src/commands/upload/submission-service/android/AndroidSubmitter.ts
@@ -1,16 +1,13 @@
-import attempt from 'lodash/attempt';
-import isError from 'lodash/isError';
 import pick from 'lodash/pick';
 import ora from 'ora';
-import axios from 'axios';
 
-import log from '../../../../log';
 import { AndroidSubmissionConfig } from './AndroidSubmissionConfig';
 import { ServiceAccountSource, getServiceAccountAsync } from './ServiceAccountSource';
 import { ArchiveSource, getArchiveUrlAsync } from '../ArchiveSource';
 import SubmissionService, { Platform, Submission, SubmissionStatus } from '../SubmissionService';
 import { AndroidPackageSource, getAndroidPackageAsync } from './AndroidPackageSource';
 import { sleep } from '../../../utils/promise';
+import { displayLogs } from '../utils/logs';
 
 export interface AndroidSubmissionOptions
   extends Pick<AndroidSubmissionConfig, 'archiveType' | 'track' | 'releaseStatus'> {
@@ -22,7 +19,7 @@ export interface AndroidSubmissionOptions
 class AndroidSubmitter {
   constructor(private options: AndroidSubmissionOptions, private verbose: boolean = false) {}
 
-  async submit() {
+  async submitAsync() {
     const submissionConfig = await this.formatSubmissionConfig();
     const scheduleSpinner = ora('Scheduling submission').start();
     const submissionId = await SubmissionService.startSubmissionAsync(
@@ -82,58 +79,5 @@ const getStatusText = (status: SubmissionStatus): string => {
     throw new Error('This should never happen');
   }
 };
-
-async function displayLogs(submission: Submission | null): Promise<void> {
-  if (!submission?.submissionInfo?.logsUrl) {
-    // no logs available, nothing to display
-    return;
-  }
-  const rawLogs = await downloadFile(submission?.submissionInfo?.logsUrl);
-  const logs = parseLogs(rawLogs);
-  log.addNewLineIfNone();
-  const prefix = log.chalk.blueBright('[logs] ');
-  for (const { level, msg } of logs) {
-    const msgWithPrefix = `${prefix}${msg}`;
-    if (level === 'error') {
-      log.error(msgWithPrefix);
-    } else if (level === 'warn') {
-      log.warn(msgWithPrefix);
-    } else {
-      log(msgWithPrefix);
-    }
-  }
-}
-
-async function downloadFile(url: string): Promise<string> {
-  const { data } = await axios.get(url);
-  return data;
-}
-
-interface Log {
-  level: 'error' | 'warn' | 'info';
-  msg: string;
-}
-
-function parseLogs(logs: string): Log[] {
-  const lines = logs.split('\n');
-  const result: Log[] = [];
-  for (const line of lines) {
-    const parsedLine = attempt(() => JSON.parse(line));
-    if (isError(parsedLine)) {
-      continue;
-    }
-    let level: Log['level'];
-    const { level: levelNumber, msg } = parsedLine;
-    if (levelNumber >= 50) {
-      level = 'error';
-    } else if (levelNumber >= 40) {
-      level = 'warn';
-    } else {
-      level = 'info';
-    }
-    result.push({ level, msg });
-  }
-  return result;
-}
 
 export default AndroidSubmitter;
